A few relatively important updates.
	*	Added a new function to the MessageSystem class: BroadcastFinished().
		This is a hook function to simply notify broadcast system members that
		have sent a message that the system has finished sending the message to
		all the specified targets. By default, this function does nothing however
		subclasses may override it as they see fit.
	*	Eliminated the Broadcast function of the plug-in manager now that the
		new message broadcast system can touch any part of the application, the
		plug-in manager no longer needs to have its own broadcast function. That
		being said, the plug-in manager is now a part of the broadcast system as well.
		It has it's own target, but doesn't do anything with received messages. It
		only sends out a few messages, none of which require a reply: plug-in loaded,
		plug-in unloaded, and B_QUIT_REQUESTED.
	*	Many printf statements were placed between "#ifdef DEBUG" and "#endif". A
		slight change to the makefile is coming (eventually) that by default will include
		"-DDEBUG" so that the printf statements will still be compiled in. You may also
		be able to change the environmental variable CFLAGS to get the same effect.
		(I haven't tested this last part, but I do believe that it will work.)


git-svn-id: svn://svn.code.sf.net/p/themis/code@219 ecaf47ba-bb13-0410-a3b6-899b003a5108

+++ trunk/themis/common/msgsystem.cpp
@@ -16,7 +16,7 @@ MessageSystem::MessageSystem(char *msg_sys_name)
 	if (MS_Name==NULL)
 		MS_Name="MessageSystem member";
 	broadcast_status_code=B_NO_ERROR;
-	_message_queue_=new BMessageQueue();
+//	_message_queue_=new BMessageQueue();
 	_msg_receiver_sem_=create_sem(0,"message_receiver_sem");
 	_ms_receiver_quit_=0;
 	_msg_receiver_running_=0;
@@ -26,12 +26,15 @@ MessageSystem::MessageSystem(char *msg_sys_name)
 
 MessageSystem::~MessageSystem()
 {
-	_ms_receiver_quit_=1;
-	release_sem(_msg_receiver_sem_);
-	delete_sem(_msg_receiver_sem_);
-	delete _message_queue_;
-	status_t stat;
-	wait_for_thread(_msg_receiver_thread_,&stat);
+	BAutolock alock(msgsyslock);
+	if (alock.IsLocked()) {
+		_ms_receiver_quit_=1;
+		release_sem(_msg_receiver_sem_);
+		delete_sem(_msg_receiver_sem_);
+//		delete _message_queue_;
+		status_t stat;
+		wait_for_thread(_msg_receiver_thread_,&stat);
+	}
 }
 int32 MessageSystem::MS_Start_Thread(void *arg) 
 {
@@ -43,7 +46,9 @@ int32 MessageSystem::_ProcessBroadcasts_(void *data)
 {
 	while (!_Quit_Thread_) {
 		if (acquire_sem(process_sem)==B_OK) {
+#ifdef DEBUG
 			printf("MessageSystem::_ProcessBroadcasts_\n");
+#endif
 			if (has_data(_ProcessThread_)) {
 //				BMessage *msg;
 				thread_id sender_thread;
@@ -79,7 +84,7 @@ int32 MessageSystem::_ProcessBroadcasts_(void *data)
 							found=true;
 //							broadcaster->broadcast_status_code|=cur->ptr->ReceiveBroadcast(msg);
 							_message_targets_++;
-								cur->ptr->_message_queue_->AddMessage(new BMessage(*msg));
+								cur->ptr->_message_queue_.AddMessage(new BMessage(*msg));
 								if (cur->ptr->_msg_receiver_running_==0) {
 									release_sem(cur->ptr->_msg_receiver_sem_);
 								}
@@ -97,7 +102,7 @@ int32 MessageSystem::_ProcessBroadcasts_(void *data)
 							if (current_target==sender_target_id) {
 								found=true;
 								broadcaster->broadcast_successful_receives++;
-								cur->ptr->_message_queue_->AddMessage(new BMessage(*msg));
+								cur->ptr->_message_queue_.AddMessage(new BMessage(*msg));
 								if (cur->ptr->_msg_receiver_running_==0) {
 									release_sem(cur->ptr->_msg_receiver_sem_);
 								}
@@ -113,7 +118,7 @@ int32 MessageSystem::_ProcessBroadcasts_(void *data)
 								found=true;
 //								printf("calling ReceiveBroadcast()\n");
 								broadcaster->broadcast_successful_receives++;
-								cur->ptr->_message_queue_->AddMessage(new BMessage(*msg));
+								cur->ptr->_message_queue_.AddMessage(new BMessage(*msg));
 								if (cur->ptr->_msg_receiver_running_==0) {
 									release_sem(cur->ptr->_msg_receiver_sem_);
 								}
@@ -137,6 +142,7 @@ int32 MessageSystem::_ProcessBroadcasts_(void *data)
 				}
 				delete msg;
 				delete container;
+				broadcaster->BroadcastFinished();
 			}
 //			printf("processing pass done.\n");
 		}
@@ -194,19 +200,19 @@ int32 MessageSystem::_ProcessMessage_(void *arg)
 			if (_ms_receiver_quit_)
 				break;
 			me->_msg_receiver_running_=1;
-			me->_message_queue_->Lock();
-			if (!me->_message_queue_->IsEmpty()) {
-				count=me->_message_queue_->CountMessages();
+			me->_message_queue_.Lock();
+			if (!me->_message_queue_.IsEmpty()) {
+				count=me->_message_queue_.CountMessages();
 			}
-			me->_message_queue_->Unlock();
+			me->_message_queue_.Unlock();
 			if (count>0) {
 				while (count!=0) {
 					if (_ms_receiver_quit_)
 						break;
-			me->_message_queue_->Lock();
-					msg=me->_message_queue_->NextMessage();
-					count=me->_message_queue_->CountMessages();
-			me->_message_queue_->Unlock();
+			me->_message_queue_.Lock();
+					msg=me->_message_queue_.NextMessage();
+					count=me->_message_queue_.CountMessages();
+			me->_message_queue_.Unlock();
 					me->ReceiveBroadcast(msg);
 					delete msg;
 					msg=NULL;
@@ -286,10 +292,15 @@ void MessageSystem::MsgSysRegister(MessageSystem *target){
 		cur->ptr=target;
 		broadcast_target_count++;
 	}
+#ifdef DEBUG
 	printf("New receiver registered; Broadcast System Members: %ld\n",broadcast_target_count);
+#endif
 }
 uint32 MessageSystem::BroadcastTarget() 
 {
 	return MS_TARGET_UNKNOWN;
 }
-		
+void MessageSystem::BroadcastFinished()
+{
+}
+
